requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
jinja2==3.1.3
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.3.2
arxiv==2.1.0
feedparser==6.0.10
toml==0.10.2